modif for sim

+++ include/buzz_utility.h
@@ -49,7 +49,7 @@ int buzz_script_done();
 
 int update_step_test();
 
-uint16_t get_robotid();
+int get_robotid();
 
 buzzvm_t get_vm();
